+++ src/protos/lightheader.h
@@ -140,9 +140,9 @@ struct FlashData {
 
     FlashData(picoproto::Message* message) :
         mode ((Mode)message->GetUInt64(1)),
-        ledcool_current (ToOptional(message->GetUInt64Array(2))),
-        ledwarm_current (ToOptional(message->GetUInt64Array(3))),
-        flash_duration (ToOptional(message->GetUInt64Array(4))),
+        ledcool_current (ToOptional(message->GetUInt32Array(2))),
+        ledwarm_current (ToOptional(message->GetUInt32Array(3))),
+        flash_duration (ToOptional(message->GetUInt32Array(4))),
         offset_duration (ToOptional(message->GetInt64Array(5)))
     {}
 };
